feat: Implement search and filtering functionality.
- Implement search bar and filtering by tag functionality.
- Introduce `usePromptList` hook for optimized prompt management.
- Add optimistic UI updates for prompt deletion.
- Refactor profile pages and improved UI components.

+++ components/PromptCard.tsx
@@ -12,13 +12,13 @@ const PromptCard: React.FC<Props> = ({
 	post,
 	handleTagClick,
 	isProfilePage,
-    onDelete,
+	onDelete,
 }) => {
 	const { data: session } = useSession() as { data: Session | null };
 	const [isModalOpen, setIsModalOpen] = useState(false);
 	const [copied, setCopied] = useState("");
-    const [deleting, setDeleting] = useState(false);
-    const router = useRouter();
+	const [deleting, setDeleting] = useState(false);
+	const router = useRouter();
 
 	const isCurrentUserPost = session?.user?.id === post.creator?._id;
 
@@ -34,44 +34,45 @@ const PromptCard: React.FC<Props> = ({
 		setIsModalOpen(true);
 	};
 
-
 	const handleConfirmDelete = async () => {
 		setIsModalOpen(false);
-        setDeleting(true);
+		setDeleting(true);
 		try {
-			if (post._id) {
-                const response = await fetch(`/api/prompt/${post._id}?id=${post._id}`, {
-                    method: "DELETE",
-                });
-                if(!response.ok){
-                    const message = `An error has occurred: ${response.status}`;
-                    throw new Error(message)
-                }
-				onDelete?.(post._id);
+			if (post._id && onDelete) {
+				onDelete(post._id, true);
+				const response = await fetch(`/api/prompt/${post._id}?id=${post._id}`, {
+					method: "DELETE",
+				});
+				if (!response.ok) {
+					const message = `An error has occurred: ${response.status}`;
+					throw new Error(message);
+				}
 			} else {
-				throw new Error("Post id is missing");
+				throw new Error("Post id is missing or onDelete not provided");
 			}
 		} catch (error) {
 			console.error("An error occurred:", error);
+			if (post._id) {
+				onDelete?.(post._id, false);
+			}
 		} finally {
-            setDeleting(false);
-        }
+			setDeleting(false);
+		}
 	};
 
-    const handleEdit = useCallback(
+	const handleEdit = useCallback(
 		(postId: string) => {
 			router.push(`/edit-prompt?id=${postId}`);
 		},
 		[router],
 	);
 
-
 	const handleCloseModal = () => {
 		setIsModalOpen(false);
 	};
 
 	return (
-		<div className="flex flex-col gap-4 p-4 shadow-md rounded-lg bg-white overflow-hidden">
+		<div className="flex flex-col gap-4 overflow-hidden rounded-lg bg-white p-4 shadow-md">
 			<Modal
 				isOpen={isModalOpen}
 				onClose={handleCloseModal}
@@ -79,50 +80,50 @@ const PromptCard: React.FC<Props> = ({
 				title="Confirm Delete"
 				message="Are you sure you want to delete this post?"
 			/>
-			<div className="flex justify-between items-start gap-5 flex-wrap">
-                {post.creator && post.creator._id ? (
-				<Link
-					href={`/profile/${post.creator._id}`}
-					className="flex-1 flex justify-start items-center gap-3 cursor-pointer"
-				>
-					<Image
-						src={post.creator?.image || "/assets/img/default-user.svg"}
-						alt="User image"
-						width={40}
-						height={40}
-						className="rounded-full object-contain"
-					/>
-					<div className="flex flex-col">
-						<h3 className="font-satoshi font-semibold text-gray-900 truncate">
-							{post.creator?.name || "Unknown User"}
-						</h3>
-						<p className="font-inter text-sm text-gray-500 break-words max-w-full">
-							{post.creator?.email || "No email"}
-						</p>
+			<div className="flex flex-wrap items-start justify-between gap-5">
+				{post.creator && post.creator._id ? (
+					<Link
+						href={`/profile/${post.creator._id}`}
+						className="flex flex-1 cursor-pointer items-center justify-start gap-3"
+					>
+						<Image
+							src={post.creator?.image || "/assets/img/default-user.svg"}
+							alt="User image"
+							width={40}
+							height={40}
+							className="rounded-full object-contain"
+						/>
+						<div className="flex flex-col">
+							<h3 className="truncate font-satoshi font-semibold text-gray-900">
+								{post.creator?.name || "Unknown User"}
+							</h3>
+							<p className="max-w-full break-words font-inter text-sm text-gray-500">
+								{post.creator?.email || "No email"}
+							</p>
+						</div>
+					</Link>
+				) : (
+					<div className="flex flex-1 items-center justify-start gap-3">
+						<Image
+							src={"/assets/img/default-user.svg"}
+							alt="User image"
+							width={40}
+							height={40}
+							className="rounded-full object-contain"
+						/>
+						<div className="flex flex-col">
+							<h3 className="truncate font-satoshi font-semibold text-gray-900">
+								{"Unknown User"}
+							</h3>
+							<p className="max-w-full break-words font-inter text-sm text-gray-500">
+								{"No email"}
+							</p>
+						</div>
 					</div>
-				</Link>
-                  ) : (
-                    <div className="flex-1 flex justify-start items-center gap-3">
-                        <Image
-						    src={"/assets/img/default-user.svg"}
-						    alt="User image"
-						    width={40}
-						    height={40}
-						    className="rounded-full object-contain"
-					    />
-					    <div className="flex flex-col">
-						    <h3 className="font-satoshi font-semibold text-gray-900 truncate">
-							    {"Unknown User"}
-						    </h3>
-						    <p className="font-inter text-sm text-gray-500 break-words max-w-full">
-							    {"No email"}
-						    </p>
-					    </div>
-                    </div>
-                )}
+				)}
 				<div
 					onClick={handleCopy}
-					className="w-7 h-7 rounded-full bg-white/10 shadow-[inset_10px_-50px_94px_0_rgb(199,199,199,0.2)] backdrop-blur flex justify-center items-center cursor-pointer"
+					className="flex h-7 w-7 cursor-pointer items-center justify-center rounded-full bg-white/10 shadow-[inset_10px_-50px_94px_0_rgb(199,199,199,0.2)] backdrop-blur"
 				>
 					<Image
 						src={copied ? "/assets/icons/tick.svg" : "/assets/icons/copy.svg"}
@@ -132,11 +133,11 @@ const PromptCard: React.FC<Props> = ({
 					/>
 				</div>
 			</div>
-			<p className="my-4 font-satoshi text-sm text-gray-700 break-words max-w-full text-ellipsis">
+			<p className="my-4 max-w-full text-ellipsis break-words font-satoshi text-sm text-gray-700">
 				{post.prompt}
 			</p>
 			<p
-				className="font-inter text-sm blue_gradient cursor-pointer truncate"
+				className="blue_gradient cursor-pointer truncate font-inter text-sm"
 				onClick={() => handleTagClick?.(post.tag)}
 			>
 				{post.tag}
@@ -145,14 +146,14 @@ const PromptCard: React.FC<Props> = ({
 				<div className="flex justify-end gap-2">
 					<button
 						onClick={() => handleEdit(post._id || "")}
-						className="px-3 py-1 bg-blue-500 text-white rounded-md"
+						className="rounded-md bg-blue-500 px-3 py-1 text-white"
 					>
 						Edit
 					</button>
 					<button
-                        onClick={handleDeleteClick}
+						onClick={handleDeleteClick}
 						disabled={deleting}
-						className="px-3 py-1 bg-red-500 text-white rounded-md"
+						className="rounded-md bg-red-500 px-3 py-1 text-white"
 					>
 						{deleting ? "Deleting..." : "Delete"}
 					</button>
